Setup a basic continuous integration workflow for a Django app
* Try running the CI with python3 instead of python command

Since the last run failed "ModuleNotFoundError: No module named 'environ'" Error which is attributed to some python venv shenanigans, we retry by running the python3 command explicitly.

* Test Changes to get the workflow running

Some more tests in order to get the workflow running.

* Change Django CI Workflow to use Poetry

Since we are using poetry in our project, we can run the CI workflow using poetry as well for now.
In later iterations, this should be changed since it is most likely not desirable to run poetry in production environments.

* Fix Caching of venv

Use the newest version of the cache action (v4).
Fix the reference to the cache in the poetry install phrase.

* Use poetry with virtualenvs

Change Version of install-poetry-action to v1.3 and add the with: create_virtualenvs: true clause.

* Switch to snok/install-poetry@v1

since the install-poetry-action is no longer developed, switch to an action that is current and still maintained.

* Fix wrong indentation

* Activate venv during test runs

* Fix wrong indentation

* Try to resolve the venv issues

* Fix missing bracket

* Further fixes

* Update django.yml

* Fix virtualenv in project during workflow run

* Add a secret key to the ENV

* Pin snok/install-poetry to full hash instead of version

CodeQL warning suggested to pin the third party action to a commit hash instead of the version number,